Space

+++ statistics.c
@@ -1,7 +1,6 @@
+#include "statistics.h"
 
-
-
-double jaf_sum(double * numbers, size_t size) {
+double jaf_sum(double * numbers, int size) {
 	double sum = 0;
 	for (int i = 0; i < size; ++i) {
 		sum += numbers[i];
